requirements.txt
absl-py==1.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
-e git+https://github.com/KaiyangZhou/Dassl.pytorch.git@6fc7902ff41870b4eac591c81de1c47c7a63edec#egg=dassl
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.8.0
flake8==3.7.9
fonttools==4.37.3
ftfy==6.1.1
future==0.18.2
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.12.0
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
grpcio==1.49.1
idna==3.4
importlib-metadata==4.12.0
isort==4.3.21
joblib==1.2.0
kiwisolver==1.4.4
littleutils==0.2.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
mccabe==0.6.1
numpy==1.23.4
oauthlib==3.2.1
ogb==1.3.4
onedrivedownloader==1.1.3
outdated==0.2.1
packaging==21.3
pandas==1.5.1
pathtools==0.1.2
Pillow==9.3.0
promise==2.3
protobuf==3.20.2
psutil==5.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
quadprog==0.1.11
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
sentry-sdk==1.9.10
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
tabulate==0.8.10
tb-nightly==2.11.0a20220929
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.13.4
wcwidth==0.2.5
Werkzeug==2.2.2
wilds==1.2.2
yacs==0.1.8
yapf==0.29.0
zipp==3.8.1
